added comments.

+++ database/update.py
@@ -1,24 +1,50 @@
 #!/usr/bin/python
 
+# * Modules
 import pandas as pd
 import json
 
+# Local Imports
 from sql_shell import connect_to_db
+
+# ? VSCode Extensions Used:
+# ?     - Better Comments
+# ?     - autoDocstring
+
+# * Quick Reference =============================================================================
+# * update_from_csv         => Updates the database from a csv calculations file.
+# * get_dashboard_variables => Gets the dashboard variables defined in dashboard_variables.json
+# * update_schema_file      => Checks the dashboard variables json file and updates the SQL
+# *                            schema with it.
+# * update_schema           => Updates the database calculations table columns given
+# *                            updates from the dashboard variables json file.
+# * =============================================================================================
 
 TEST_FILE = R"C:\dev\Gerontechnology\data wrangling\data\output\Week 14, 2022.csv"
 
+def update_from_csv(filename: str):
+    """Updates the database from a csv calculations file.
 
-def update_from_csv(filename):
+    Args:
+        filename (str): The csv file to reference new variable calculations from.
+    """
     calculations_table = pd.read_csv(filename)
     print(calculations_table)
 
 def get_dashboard_variables():
+    """Gets the dashboard variables defined in dashboard_variables.json
+
+    Returns:
+        List(dict): A list of dictionaries that is [{"name":str, "type":str}]
+    """
     with open("dashboard_variables.json") as json_file:
         data = json.load(json_file)
         return data['Variables']
 
 
 def update_schema_file():
+    """ Updates the schema sql file that is referenced by create_db()
+    """
     variable_schema = get_dashboard_variables()
     variables = ""
 
@@ -54,7 +80,14 @@ CREATE TABLE IF NOT EXISTS Calculations
     with open("EMMA_variables_schema.sql", "w") as fd:
         fd.write(sql_schema)
 
-def update_schema():
+def update_schema(force_delete=False):
+    """ Updates the database calculations table columns given
+        updates from the dashboard variables json file.
+
+    Args:
+        force_delete (bool, optional): When true, will drop columns from the calculation table
+        if those columns are no longer defined in dashboard variables. Defaults to False, to avoid accidental deletions.
+    """
     update_schema_file()
 
     variables = get_dashboard_variables()
@@ -78,13 +111,13 @@ def update_schema():
     for variable in extra_variables:
         sql_str = "ALTER TABLE Calculations Add {} {};".format(variable["name"], variable["type"])
         cursor.execute(sql_str)
-
-    for variable in variables_to_drop:
-        sql_str = "ALTER TABLE Calculations DROP COLUMN {};".format(variable)
-        cursor.execute(sql_str)
+    if(force_delete):
+        for variable in variables_to_drop:
+            sql_str = "ALTER TABLE Calculations DROP COLUMN {};".format(variable)
+            cursor.execute(sql_str)
     
     cxn.commit()
 
 if __name__ == "__main__":
     #update_from_csv(TEST_FILE)
-    update_schema()
+    update_schema(force_delete=True)
